[types] rework content-like types
