+++ polarity/types/base.py
@@ -1,8 +1,5 @@
 import json
-from dataclasses import asdict, dataclass, field
-from datetime import datetime
-from time import sleep
-from typing import List
+from dataclasses import asdict
 
 
 class MetaMediaType(type):
@@ -28,67 +25,3 @@ class MediaType(metaclass=MetaMediaType):
         :return: JSON string
         """
         return json.dumps(asdict(self), indent=indentation)
-
-
-@dataclass
-class Content(MediaType, metaclass=MetaMediaType):
-    title: str
-    id: str
-    synopsis: str = ""
-    # TODO: better default for date
-    date: datetime = field(default=None)
-    images: list = field(default_factory=list)
-    streams: list = field(default_factory=list)
-    skip_download = None
-
-
-@dataclass
-class ContentContainer(MediaType, metaclass=MetaMediaType):
-    title: str
-    id: str
-    images: list = field(default_factory=list)
-    content: list[Content] = field(init=False, default_factory=list)
-    _extractor: str = field(init=False, default=None)
-    # True if all requested contents have been extracted, False if not
-    _extracted = False
-
-    def get_all_content(self, pop=False) -> List[Content]:
-        """
-        :param pop: (fakely) removes content from the list
-        :returns: List with extracted content
-        """
-
-        everything = []
-
-        for content in self.content:
-            if isinstance(content, ContentContainer):
-                # iterate though subcontainer contents
-                everything.extend(content.get_all_content())
-            if pop:
-                if hasattr(content, "_popped"):
-                    # if content has been popped skip to next
-                    continue
-                content._popped = None
-            everything.append(content)
-
-        return everything
-
-    def get_content_by_id(self, content_id: str) -> Content:
-        """
-        Get a Content or ContentContainer object by it's identifier
-
-        :param content_id: Content identifier to look for
-        :return: If exists returns a Content or ContentContainer object, else None
-        """
-        for content in self.content:
-            if content.id == content_id:
-                return content
-            if isinstance(content, ContentContainer):
-                _content = content.get_content_by_id(content_id)
-                if _content:
-                    return _content
-
-    def halt_until_extracted(self):
-        """Sleep until extraction has finished, useful for scripting"""
-        while not self._extracted:
-            sleep(0.1)
